1.10.5
- As funçoes passadas para arquivos próprios (para diminuir a poluição no index)
- Token agora é obtido de uma variável ambiental
