+++ .gitignore
@@ -1,2 +1,2 @@
-Token.txt
+.env
 AWS key
