+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jacadilo-bot",
-  "version": "1.10.4",
+  "version": "1.10.5",
   "description": "Personal Discord Bot",
   "main": "index.js",
   "scripts": {
@@ -12,6 +12,7 @@
   "dependencies": {
     "cheerio": "^1.0.0-rc.3",
     "discord.js": "^11.6.4",
+    "dotenv": "^8.2.0",
     "node-schedule": "^1.3.2",
     "opusscript": "0.0.7",
     "quick.db": "^7.1.1",
